Add ability to compute local search distance
This allows us to drastically speed up the time it takes
to determine what seed each pixel is closest to.
By using essentially what is a heuristic we allow for
extremely large images to be computed quickly.

There is the possibility (I think) that we determine the
incorrect seed color for any given pixel, but I
haven't put a ton of thought into that and have never
actually seen that happen.

Without this we are limited to about 1 million pixels.